python/requirements.txt
grpcio>=1.62.0
grpcio-reflection>=1.62.0
protobuf>=4.25.0
prompt_toolkit>=3.0.43
pygments>=2.17.0
